Initial commit
gitignore:
It may be a good idea to ensure we don't commit potentially large
objects to our repository which we don't require in order to build the
snap.

LICENSE:
This is the license of all the files in this repository (unless
otherwise specified). This license information should show up in the
snapcraft.yaml of our snap.

snap/snapcraft.yaml:
The trivial snap. It doesn't do anything.
This will soon change.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>