snapcraft.yaml: add initial apps and parts
Here the gotop binary is built and the app the snap is going to provide
is defined.

Notice the use of the nil plugin here. If you inspect the go plugin:
https://github.com/canonical/snapcraft/blob/main/snapcraft_legacy/plugins/v2/go.py

You'll see that it offers a very opinionated way of building go
packages. Indeed, such a method doesn't appear to be directly supported
by the way the gotop repository is structured:
https://github.com/xxxserxxx/gotop?tab=readme-ov-file#building

As such, the build will have to be handled manually.

Note very importantly that all overrides are shell scripts run by the
dash shell. The dash shell is less featured than something you may be
more used to, like bash. Specifically, dash is a more strictly POSIX
compliant shell. As a reference, see:
https://pubs.opengroup.org/onlinepubs/9699919799/

This doesn't however stop you from using bash or even coding an override
in an entirely different language like C or python.

Instead of writing something like:

  override-build: |
    <shell script>

You could instead write up a bash script or compile some C code and do:

  override-build: |
    bash path/to/bash/script
    # or
    path/to/C/executable

Indeed, if GCC is available in your build environment, you could create
a part that compiles that code and makes it available for this part to
use DURING the snap build!

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>