snapcraft.yaml: add missing build package
In order to build gotop from source, particular build-time dependencies
are required. For an explanation on how these are handled, see here:
https://snapcraft.io/docs/build-and-staging-dependencies

Conveniently, gotop only requires the go compiler to be built.

Resolves this build error:

dilyn@Ares:~/git/gotop -> snapcraft
'override-build' in part 'gotop' failed with code 127.
Recommended resolution: Review the scriptlet and make sure it's correct.
Failed to execute snapcraft in instance.
Full execution log: '/home/dilyn/.local/state/snapcraft/log/snapcraft-20240625-000056.346799.log'

We get more illuminating information by searching the mentioned log
file for when we start building the gotop part, as the error message
indicates a failure in our override-build step for the gotop part:

2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.208 Building gotop
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.208 execute action gotop:Action(part_name='gotop', step=Step.BUILD, action_type=ActionType.RUN, reason=None, project_vars=None, properties=ActionProperties(changed_files=None, changed_dirs=None))
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.208 load state file: /root/parts/gotop/state/pull
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.212 remove directory /root/parts/gotop/build
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.245 :: + go build -o gotop ./cmd/gotop
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.246 :: /bin/bash: line 44: go: command not found
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.442 'override-build' in part 'gotop' failed with code 127.
2024-06-25 00:02:08.927 :: 2024-06-25 00:02:08.445 Traceback (most recent call last):

Note that instead of investigating the build log to try and find the
error message, we can build with the --verbose flag to get better
invisibility into the build as it is happening.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>