README: add installation instructions
There are several ways to install a snap!

If the snap package does not have an accompanying assertion file (e.g.
it did not come from the store), then it must be installed with the
`--dangerous` flag.
If the snap package specifies `confinement: devmode`, it must installed
with the `--devmode` flag. `--devmode` supercedes the `--dangerous` flag.

Note that a snap with `confinement: devmode` can be installed as if it
had `confinement: strict` by using the `--jailmode` flag.

For more information on install modes for snaps, see here:
https://snapcraft.io/docs/install-modes

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>