snapcraft.yaml: use organize keyword to simplify
The organize keyword is useful for moving directories and files around
to avoid overly long overrides in your parts.

In this instance, we change the directory we install the gotop binary
to and move it to usr/bin in an organize, likewise for the blkid binary
(moving from usr/sbin to usr/bin).

Note that importantly, an organize takes effect at the end of a part's
build step and before the stage step. This means that any file filtering
we do in the stage keyword will have to reflect any changes made in an
organize step.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>