snapcraft.yaml: snap version is gotop version
The version of a snap can be set programmatically. This can be used to
a variety of advantages, from setting the version to a particular commit
to involving the date or any number of other factors.

The character limit on a version string is 32 characters, and can
include alphanumerics, hyphens, periods.

Here, the version is set in the gotop binary itself as suggested by
the maintainer to inform users what version was built and when, and the
version of the snap is set from a subset of this information.

Note that because the method involves inspecting git information and the
release tarballs do not include this information, the source itself must
change. If we wanted to use the release tarball, we would have to change
how we determine the version.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>