snapcraft.yaml: improve binary compression
It may prove useful to remove things like debug information from
final binaries to reduce the size. This is especially true for golang
applications as they are statically linked by default, and as a result
quite sizeable.

Note that by doing this, we may lose useful information -- in the
future, it may prove useful to only use these flags for production
builds and keep the debug symbols in for our development and testing
work.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>