snapcraft.yaml: support cross-building
In addition to supporting building one snap for many different
architectures, building a snap for one arch while on a different one is
also possible.

snapcraft exposes a handful of useful environment variables which can be
used while building, including facts about the architecture the current
build is for. Refer to here:
https://snapcraft.io/docs/parts-environment-variables

Handily, golang easily supports cross-compilation -- we just need
to specify a particular GOARCH in the case where we are actually
cross-compiling our executable. We can readily achieve this by
leveraging the environment variables exposed by snapcraft; indeed, we
could hardcode values in some places as well. Note what we do in the
newly added build-environment section of our gotop part, which sets the
GOARCH variable depending on the target architecture of the build. When
natively building (which our platforms section also allows), GOARCH
is set to the host architecture anyways so no work is required in that
case.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>