snapcraft.yaml: fix builds on riscv64, s390x
Like ARM64, both RISC-V 64-bit and s390x architectures do not support
the dup2 syscall, but instead support dup3.

This can be remedied by extending the solution for ARM64 systems to
the others.

With a little target-arch and golang knowledge, a solution readily
exists. Indeed, patches already exist accomplishing this on Alpine:
https://git.alpinelinux.org/aports/tree/community/gotop/0001-logging-add-support-for-riscv64.patch
As well as a pending PR for upstream:
https://github.com/xxxserxxx/gotop/pull/218

There are many ways to handle this. In this case, the patch has been
fetched and added to the repository in a new patches/ directory.
snapcraft makes the project directory of our snap available to our
snap, which means we can refer to this patch file by leveraging the
${CRAFT_PROJECT_DIR} variable.

Note that the patching is performed in a new override-pull step, as the
pull step won't be reran on build failures. This way, we don't try to
patch our source multiple times when doing iterative builds. If this
patch were to be applied in the override-build step, it's possible that
iterative builds would fail because the patch has already been applied.

In short, doing the patching in an override-pull gives a soft guarantee
that the patch isn't applied multiple times as, if the source or
override-pull step changes, the pull step will be redone giving us a
clean src directory anyways.

Resolves this build error:

: # github.com/xxxserxxx/gotop/v4/logging
:: logging/logging_other.go:11:10: undefined: syscall.Dup2
'override-build' in part 'gotop' failed with code 1.

Note that the patch which has been added to this repository assumes the
license of the upstream project.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>